do not promote api version param to client if service is not versioned
